% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_cli.R
\name{display_explanation}
\alias{display_explanation}
\title{Filter how to process HTML codes}
\usage{
display_explanation(text, use_html = shiny::isRunning())
}
\description{
Filter how to process HTML codes
}
