% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_cli.R
\name{filter_words}
\alias{filter_words}
\title{Filter words based on user choices}
\usage{
filter_words(
  wordbank,
  column_filters = list(),
  keyword_filter = NULL,
  rank_limit = NULL
)
}
\description{
Filter words based on user choices
}
