% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_cli.R
\name{load_user_scores}
\alias{load_user_scores}
\title{Load previous scores to update memory}
\usage{
load_user_scores(
  wordbank,
  username = NULL,
  save_dir = "user_data",
  interactive_mode = TRUE
)
}
\description{
Load previous scores to update memory
}
