% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_cli.R
\name{run_quiz_cli}
\alias{run_quiz_cli}
\title{Run the quiz}
\usage{
run_quiz_cli(
  wordbank,
  keyword_filter = NULL,
  n_questions = 10,
  rank_limit = NULL,
  interactive_mode = TRUE,
  selection_mode = "random",
  start_rank = 1
)
}
\description{
Run the quiz
}
